requirements.txt
flasgger==0.9.5
Flask==2.2.2
numpy==1.23.2
pandas==1.4.3
Pillow==9.2.0
streamlit==1.12.2
